refactor(v1): keep AgentConfig in agent.py; drop the runtime migration guard
AgentConfig/TimeoutConfig return to agent.py (config next to its class,
like every other plugin). The trace record keeps its typed embed without
an import cycle by base-annotating `AgentInfo.config` and narrowing to
`AgentConfig` in a validator whose lazy import runs at validation time,
not import time. Also removes HarnessConfig's harness-level `runtime`
pointer validator — the stray key now fails as a plain unknown field.
