+++ verifiers/v1/agent.py
@@ -10,16 +10,20 @@ from collections.abc import Callable, Iterator, Mapping
 from contextlib import asynccontextmanager, nullcontext
 from typing import AsyncIterator
 
+from pydantic import SerializeAsAny, model_validator
+from pydantic_config import BaseConfig
+
 from verifiers.v1.clients import (
     Client,
+    ClientConfig,
     EvalClientConfig,
     ModelContext,
     resolve_client,
 )
-from verifiers.v1.harness import AgentConfig, TimeoutConfig
+from verifiers.v1.harness import HarnessConfig
 from verifiers.v1.interception import Interception, InterceptionServer
 from verifiers.v1.mcp import SharedToolServer
-from verifiers.v1.retries import backoff, trace_should_retry
+from verifiers.v1.retries import RetryConfig, backoff, trace_should_retry
 from verifiers.v1.rollout import RolloutRun
 from verifiers.v1.runtimes import (
     DockerConfig,
@@ -32,7 +36,7 @@ from verifiers.v1.runtimes import (
 from verifiers.v1.session import RolloutLimits
 from verifiers.v1.task import Task
 from verifiers.v1.trace import Trace
-from verifiers.v1.types import Sampling
+from verifiers.v1.types import Sampling, SamplingConfig
 from verifiers.v1.utils.compile import (
     cap_remote_harness_timeout,
     resolve_runtime_config,
@@ -42,6 +46,57 @@ from verifiers.v1.utils.compile import (
 __all__ = ["Agent", "AgentConfig", "Agents", "TimeoutConfig", "make_agent"]
 
 logger = logging.getLogger(__name__)
+
+
+class TimeoutConfig(BaseConfig):
+    """Per-agent wall-clock timeouts per rollout stage, in seconds (None = no
+    limit); each stage falls back to the task's own `TaskTimeout` when unset."""
+
+    setup: float | None = None  # one shared budget: task setup + provisioning
+    rollout: float | None = None
+    finalize: float | None = None
+    scoring: float | None = None
+
+
+class AgentConfig(BaseConfig):
+    """One env agent: who plays it, and its per-run caps. It pins only what
+    makes it a different actor; everything unpinned falls back — the model context
+    to the run's own, the harness to the taskset's default."""
+
+    harness: SerializeAsAny[HarnessConfig] | None = None
+    """The agent's program (None = the taskset's default harness)."""
+    runtime: RuntimeConfig = SubprocessConfig()
+    """Runtime for the harness program — the policy each run provisions its box
+    from; tool servers choose their placement separately."""
+    model: str | None = None
+    """Model id (None = the run's model, i.e. the policy under evaluation/training)."""
+    client: ClientConfig | None = None
+    """Endpoint override (None = the run's client)."""
+    sampling: SamplingConfig | None = None
+    """Sampling override (None = the run's sampling)."""
+    timeout: TimeoutConfig = TimeoutConfig()
+    retries: RetryConfig = RetryConfig()
+    """Whole-run retries: rerun this agent's rollout while its trace ends with a
+    retryable error (never into a borrowed box)."""
+    max_turns: int | None = None
+    """Max model turns per run (None = no limit). Framework-enforced (the
+    interception server refuses turns past it), so it applies to any harness."""
+    max_input_tokens: int | None = None
+    max_output_tokens: int | None = None
+    max_total_tokens: int | None = None
+    """Token caps per run (None = no limit); framework-enforced between turns."""
+
+    @model_validator(mode="before")
+    @classmethod
+    def _resolve_harness(cls, data):
+        """Narrow a pinned `harness` to its concrete config type by `id` (absent
+        stays None = the taskset's default). The lazy import keeps class-body
+        `AgentConfig()` defaults constructible while this module initializes."""
+        if isinstance(data, dict) and data.get("harness") is not None:
+            from verifiers.v1.loaders import harness_config_type, narrow_plugin_field
+
+            narrow_plugin_field(data, "harness", harness_config_type, "bash")
+        return data
 
 
 def _check_borrowed_placement(
