+++ verifiers/v1/trace.py
@@ -8,16 +8,17 @@ from collections.abc import Mapping
 from typing import TYPE_CHECKING, Annotated, Any, Generic, Literal
 
 import numpy as np
-from pydantic import Field, PrivateAttr
+from pydantic import Field, PrivateAttr, SerializeAsAny, field_validator
+from pydantic_config import BaseConfig
 from renderers.base import MultiModalData
 
 if TYPE_CHECKING:
+    from verifiers.v1.agent import AgentConfig
     from verifiers.v1.judge import JudgeResponse
 
 from verifiers.v1 import graph
 from verifiers.v1.errors import ProviderError
 from verifiers.v1.graph import MessageNode
-from verifiers.v1.harness import AgentConfig
 from verifiers.v1.runtimes import RuntimeInfo
 from verifiers.v1.state import State, StateT
 from verifiers.v1.task import DataT, WireTaskData
@@ -310,12 +311,14 @@ class AgentInfo(StrictBaseModel):
     `AgentConfig` it ran as, the box it ran in, and its standing in the episode;
     the `Episode` envelope links siblings."""
 
-    config: AgentConfig
-    """The agent's resolved config, lossless — the exact value that rebuilds this
-    agent (`Agent(trace.agent.config)`): harness, runtime policy, model, sampling,
-    caps. Re-validating a record narrows `config.harness` by its id, so a typed
-    read needs the harness package importable (a custom harness's extra knobs
-    serialize and read back typed)."""
+    config: SerializeAsAny[BaseConfig]
+    """The agent's resolved `AgentConfig`, lossless — the exact value that rebuilds
+    this agent (`Agent(trace.agent.config)`): harness, runtime policy, model,
+    sampling, caps. Base-annotated and narrowed by the validator below, so the
+    trace record stays below `agent.py` in the import graph. Re-validating a
+    record narrows `config.harness` by its id, so a typed read needs the harness
+    package importable (a custom harness's extra knobs serialize and read back
+    typed)."""
     runtime: RuntimeInfo | None = None
     """The box the rollout ran in — `config.runtime` resolved for the task, plus
     the provisioned resource ID; None until provisioning."""
@@ -326,6 +329,17 @@ class AgentInfo(StrictBaseModel):
     trainable: bool = True
     """Whether this trace's tokens are training data for the run's policy. An env's
     `setup()` marks fixed-model agents (a frozen judge, a pinned user sim) untrainable."""
+
+    @field_validator("config", mode="before")
+    @classmethod
+    def _narrow_config(cls, value) -> "AgentConfig":
+        """Narrow `config` to a real `AgentConfig`. The lazy import runs at
+        validation time, keeping this module importable below `agent.py`."""
+        from verifiers.v1.agent import AgentConfig
+
+        if isinstance(value, AgentConfig):
+            return value
+        return AgentConfig.model_validate(value)
 
 
 class TraceTask(StrictBaseModel, Generic[DataT]):
